Wrote lexer
Hopefully. I think I did, anyways.

The symbol lexing will probably need to be broken off into its own
sub-lexer like the parenthesis group one, as will any token more
complicated than the hashmark token lexing.

Next comes the parser.

The actual function should be straightforward: hook_print on the
"Channel Message" text event for now, and figure out a better way to
capture different types of messages in the future.

Also, the lexer doesn't enforce wrapping Lua patterns in parenthesis,
or making sure the # is not floating, and is with something,
like #irc or #(irc%-.*).

The parser is supposed to do this anyways, and I'm considering not
enforcing anything, and simply letting the lexer and parser dole out
errors to those who don't follow the grammar.

For now, # can be dangling in free space, and
Lua patterns without spaces are fine.

For example:
/SSOUND freenode #irc sound H(:\sound.wav match .*

is perfectly valid, currently.