Really not working
Tried to add settings getter and setter.

The Programming in Lua book has some really cool recipes
for storing and retrieving data, using loadstring.

I don't think I need that, though:

I wrote a serialization function that escapes and wraps whaetever it's
given in [[ ]], as it appears completely reversible
using gsub.

I can just use the settings_prefix, and then concat that with the server name
and channel name, both escaped and [[ ]]-wrapped,
and use that as the settings_key,
then simply concat sound and match with a newline, escaping existing
newlines if I want to.

This would greatly simplify and speed up the code, I think.