BROKEN: Need to make splay_on_matching_message work
Need to add a global settings table so every time a setting needs
to be checked, it can just be looked up in a table.
store_settings needs to update this table
setup needs to initialize this table
retrieve_settings should only access the disk if this table returns nil

retrieve_settings should return an iterator, and all calls need to be
updated for this

hooking the text event "Channel Message" misses private messages
and /ACTION messages and such

Each server+channel combo has to be able to store multiple
match patterns.