Partially working
There was a loop error, as get_settings hadn't been updated to deal with the
iterator returned by retireve_settings

Cannot delete a setting, since you can only specify the new setting to
be set, and not the old setting to overwrite.

Need to revisit actions setup, and add some actions.