Begin refactor
Got things working, tried to add features, and broke it, instead of
committing.

So now I'm planning to refactor the parsing functions called by
inner_parse() so that the everything works a bit smoother, in the call
chaining.

I want each action to have a corresponsing action function, and
then the default action can be show_or_add_or_delete(),
and that can simply collect tokens until it meets a particular state,
where it passes control to the appropriate function.

I was worried that setting hexchatpluginprefs keys to nil wouldn't
delete keys from the file, and would instead do:

key =

But it works appropriately, even so far as to delete the entire settings
file if all keys are removed.

If ithadn't, it would have been difficult to have made /delete and
/deleteall atomic, requiring file io from Lua.