/show_or_add_or_delete appears to work
The small, happy path test run in the comment at the head of the file
caught a but in delete_action()

I need to have catching else conditions for every if in these forever
looping functions.

They also catch where I missed a case, or state.

/parse doesn't return debug_on back to its original setting.

It could connect a hook timer on a very short delay that has a closure
for a callback, that resets it and then unhooks itself, so that the
setting is always reset, even if the command fails with a Lua backtrace.

Next, sound file validation.