Begin adding bells and whistles
test.lua works, build.lua is not written

The main file is probably going to be SpecialSounds.lua, and will need
to have a conditional at the end testing if the hexchat table is
available.

If it is, it should run the setup() function, and if not, it should:

- check if the variable tests is true and run tests
- check if the variable build is true and run build

That way, the lua interpreter can be used for running tests and
building, like:

```sh
lua -e "tests=true build=true" SpecialSounds.lua
```

So that tests can be run without requiring a build to be done.

I think a version bump should also be able to be optionally specified,
so:

- tests
- build
- bumpver

```sh
lua -e "tests=1 build=1 bumpver=1" SpecialSounds.lua
```

My dream would be to have this project autotest and build and set
releases on pushes to master, just like a real project.