Broken! Half-way through factoring out print_error
build.lua is also not written, but it did prompt considerations about
the general architecture of the system.

I might just use simple print_*() functions inside the module to get the
build and tests running, then work on factoring out the print functions.

Cool part is, as long as I don't require the print.lua file, I don't
have to include it in the build yet.