building happens as expected
Built file is UNTESTED

It's probably be a good idea to eventually have the test.lua be included
in the file so that tests can be run on the file as a whole, to make
sure throwing all the files together doesn't create a problem.

It would also be very sweet if the setup() function in main.lua could
register a /tests action if tests have been built in, to run the full
test sweet in HexChat, allowing to verify if everything is working as
expected, and to test for errors.

Files get mushed together! Yay!

I'm hoping CI/CD can submit build artifacts so that ugly mush doesn't
have to be committed to the repository.