Trying to make print output available to tests
Not fully integrated into test runners or test summarization, as a lot
of variable renaming was done, for both integrating print_output, and
changing the functions that record printing.

Changed emit.record to record the type of emit as a tag, so the array of
records is now an array where each element is a pair of tag and message:

local print_records_example = {
  {"print", "Hi! Testing"},
  {"info",  "Here's some info"},
  {"err",   "Uhoh, an error!"},
  {"write", "Wrote it!"},
}

Moved unused read-only table generator to header.lua, along with
very-used compare_tables. Both don't play well with tables that have
cyclical table references.

Calling the function returned by emit.record_output/record_prints no
longer turns printing back on, so nested calls can retrieve their output
without turning on printing globally again. I can't believe this wasn't
caught before now.

The above is implemented with reference counting that feels flaky.

Moved test cases and test functions to tests.lua, to separate them from
the test runners. This means the skip_wip_tests global flag is checked
everytime that file is loaded as a chunk, which I'm hoping means once
for each run of the program; otherwise WIP test cases will be duplicated.

Got rid of test.add_result() as it just was not something that was
needed, and was poorly implemented to begin with. I think it would have
been better to arrange the results table in such a way that keys could
be used, but that would require a much larger rewrite. An easier rewrite
that also would have needed to have been done, had the function still
been in use, would be to change the long list of function parameters
into a single table, whose keys could exist or not, as a way of passing
arguments by name to a function.

OR! use varargs! {...} duh!

In general, there a few function calls here and there that are made of
functions that could return the standard handleable error or:

function ()
  return nil, "Error message"
end

That are being called in places where the error should be raised
immediately, as it can't be handled, like when the call is made inside a
test function.

These calls should be wrapped in assert(). I didn't think of this until
now.