GitHub Actions workflow feature complete
While I think I have a complete set of GitHub Actions, I'm still iffy on
how they're triggered.

I wanted to keep the part of the process that resulted in a merge commit
to master, signed by me, and a tag pointing to it, marking a new
release, also signed by me.

I do intend to figure out a way to push changes made to the repository
during a workflow from the Actions runner, that includes commit and tag
signatures, but I'm putting that on the backburner.

Using the GitHub Python package to make commits using the GitHub API
seems promising, as those would be signed with GitHub's key:
https://github.community/t5/GitHub-Actions/how-does-one-commit-from-an-action/m-p/31870/highlight/true#M921