Reordered actions; Set PATH the right way
I don't feel confident in the way that the cache key is created fr the
actions/cache action, as I'm not sure where it's searching for the
version files.

This also comes fro the fact that the run for the previous commit seems
to have succeeded, with the cache saving the file, but also has a
message in the GitHub interface notifying of an error:
https://github.com/2ndBillingCycle/specialsounds/runs/459199092?check_suite_focus=true#step:25:1

I reordered the actions to "unfix" the problem, and added a debug
message and key to my secrets to try to print the path of where the
command is being run BEFORE the checkout action, since running the
checkout action first seemed to fix the issue.

Using the instructions from here:
https://help.github.com/en/actions/reference/development-tools-for-github-actions#set-a-debug-message-debug

On the above page, it contains instructions for manipulating $PATH, so I
changed the way I it's set in the current workflow:
https://help.github.com/en/actions/reference/development-tools-for-github-actions#add-a-system-path-add-path