build.lua and test.lua now behave like scripts, and are called like:
lua build.lua test build
lua test.lua

One of the disadvantages of not reading too much about how Lua is used
is that I didn't bother to look up how Lua scripts are written. Still
haven't, but using Lua's built in facilities for capturing arguments
passed on the command line seems more sane than setting global variables
as part of an expression run before the script is executed.

workflow.yml was updated to reflect these changes.

I noticed in one of the logs for a GitHub Action run that the step that
determines whether or not DO_RELEASE is set to true or false was
returning an error, despite not even being run due to a pull_request
event.

So that step was changed to never return an fail, and set DO_RELEASE to
"true" or "false". This does mean I don't technically need to declare
the environment variable at the beginning of the file, but I like being
a bit more explicit about the state management.

NOT having that variable declared may allow the use of
if: env.DO_RELEASE
without having to test its value, which would be nice...

Also, made the step that checks the program version in header.lua set an
error message for the GitHub Actions, so that it shows more than "step
failed".

Lastly, fixed a few typos, made some parts use Lua idioms more
strongly, and removed a spurious file.