Changed case tables structure; test.lua needs to be updated to match
This change is meant to simpllify a lot of the logic in the testing
framework, as it really doesn't need to be as complicated as it is.

Being more strict about the structure of the case tables will help
reduce the amount of if...else in test.lua.

What I'd also like to do is simplify the results array, and make it into
a table where the keys are the functions under test, and the value for
each key is an array that contains the return values from xpcall(), and
the inputs/outputs from the case table.

I think performing the test analysis in the summarization function is a
better place to contain this complexity, than within the test runners.