Added notes about bundling and parsing
In short:
I want to keep this as a single file download.
I also want to use an external library for parsing: LPEG.
This library is written in C.
My current method for amalgamating all the files may not work for that.

This will almost certainly mean that each platform this is meant to be
used on will require its own build of the C libraries.