I don't think I'm testing the correct way
I'm rewriting the test summarization, but I'm not exactly sure how I
want it to look, so I'm writing the tests out so I can get a feel for
how it'll be used, how I'd like it to react to different scenarios, and
hopefully that'll help me see the patterns that I can best factor out.

Also, fixed some typos.

I think what'd be best is to write a full test suite for the testing
framework. I was avoiding doing that, mainly because I wasn't sure how,
but I guess that just means it's not modular enough, in the right ways,
to test easily.