Tests added; testing halted
I think I was going about testing in the wrong way. Or rather, I wasn't
separating interfaces along the best boundaries:

Currently, the ./test.lua is supposed to be the "testing framework", but
includes the test runner, and ./tests.lua is supposed to have the tests,
but also includes helper functions that should probably be part of the
test framework.

I think redoing the test framework to double down on functions that help
assist semantic expectations, like "given this input, fails with this
error" or "all of these inputs should cause this function to return
true", would be a great benefit.

Additionally, moving a lot of the testing machinery, like the case table
parsing to the actual tests file would probably help.

I'm thinking that the testing framework functions can add their results
to a results table internal to the testing framework, which also
provides a test summarization function, or an "end tests" routine that
does the test summarization.

I think this would allow for a more clear separation, where the case
tables are defined with the tests, along with the machinery that feeds
them through the testing framework's functions, that then collect the
results, and offer facilites for processing them.

I might even split out the test runner to its own file. I've avoided
calling something a testing_framework.lua, mostly because I don't want
to encourage people to think of it as on par with busted and such.

They all are testing frameworks, in the same sense that a pile of
branches and a house are both shelters.