Start adding build system
build system is in Python mainly to make Windows+Linux easier
It's also a scripting language supported by GitHub Actions, so the
script can be fed into an action.

Hopefully this will simplify the build system, and make it so it's
possible to run it locally, or at least make it easier to figure out
what the build system is doing for someone first looking at it.

It can be modified later to be in Lua, if it's desireable to be 100%
Lua.

Also, I'm going to next work on making sure amalg can bundle:

- busted
- LPeg
- Any other dependencies

So that the testing frameowork can be switched to busted, and the parser
to LPeg.